Week 2 - Optimized
The optimized version of Week 2's assignment with addition of precalculated values and changing some of the objects' scopes to prevent unnecessary multiple initialization.

Also changed the minimal value for t to fix black pixelation of objects and the DoesHit() function, which now correctly uses the overloaded HitTest functions.

The camera now has getters and setters for the field of view to prevent unnecessary calculations each frame. Also added a field of view adjustment feature.

Fixed some of the naming conventions as well as the HitTest functions to work properly with the ignoreHitRecord parameter.